+++ src/Commands/Base/PnPSharePointCmdlet.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Management.Automation;
 using System.Net.Http;
 using System.Threading;
 using Microsoft.Online.SharePoint.TenantAdministration;
@@ -45,7 +44,7 @@ namespace PnP.PowerShell.Commands
                     {
                         var resourceUri = new Uri(Connection.Url);
                         var defaultResource = $"{resourceUri.Scheme}://{resourceUri.Authority}";
-                        return TokenHandler.GetManagedIdentityTokenAsync(this, HttpClient, defaultResource, Connection.UserAssignedManagedIdentityObjectId, Connection.UserAssignedManagedIdentityClientId).GetAwaiter().GetResult();
+                        return TokenHandler.GetManagedIdentityTokenAsync(this, HttpClient, defaultResource, Connection.UserAssignedManagedIdentityObjectId, Connection.UserAssignedManagedIdentityClientId, Connection.UserAssignedManagedIdentityAzureResourceId).GetAwaiter().GetResult();
                     }
                     else
                     {